core_timing: Split off utility functions into core_timing_util
